requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-telegram-bot==20.3
python-dotenv==1.0.0
sqlalchemy==2.0.15